harry-insta/requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
black==19.10b0
boto3==1.14.12
botocore==1.17.12
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dj-database-url==0.5.0
Django==2.1.15
django-appconf==1.0.4
django-bootstrap-pagination==1.7.1
django-bootstrap4==2.1.1
django-cleanup==5.0.0
django-extensions==2.2.9
django-heroku==0.3.1
django-imagekit==4.0.2
django-otp==0.9.3
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
jmespath==0.10.0
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.1.2
prompt-toolkit==3.0.5
psycopg2==2.7.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
qrcode==6.1
regex==2020.6.8
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.5
whitenoise==5.1.0
zipp==3.1.0
